Decompose SynthChat generator God class into focused modules
Break SynthChatGenerator (3,384→1,635 lines) and run.py (1,042→220 lines)
into 14 focused modules following the 9-step decomposition roadmap:

Extracted modules:
- template_utils.py: template rendering utilities
- targets.py: target spec normalization (DRY fix)
- workspace/: environment rendering (renderer, sections, fixture_helpers)
- schemas/: JSON schemas (environment, tool response)
- labeling.py: metadata label classification
- parsing.py: response parsing and normalization
- llm/: LLM client pool management
- review.py: stage review and judge templates
- agentic/: episode generation and turn management
- modes/: CLI mode handlers (generate, improve, validate)
- parallel/: worker pool and parallel execution
- result_writer.py: streaming output management

All 43 existing tests pass. Backward-compat re-exports preserved.
