+++ SynthChat/agentic/__init__.py
@@ -1,0 +1,7 @@
+"""SynthChat Agentic - Multi-turn agentic episode generation.
+
+Location: SynthChat/agentic/__init__.py
+Purpose: Subpackage for generating multi-turn agentic rollouts using the shared
+         environment episode runner, turn judges, and loop response callbacks.
+Usage: from SynthChat.agentic.episode import generate_agentic_episode
+"""
