+++ SynthChat/modes/__init__.py
@@ -1,0 +1,6 @@
+"""SynthChat Modes - CLI mode implementations (generate, improve, validate).
+
+Location: SynthChat/modes/__init__.py
+Purpose: Subpackage for the three CLI modes of SynthChat.
+Usage: from SynthChat.modes.generate import generate_mode
+"""
