+++ SynthChat/parallel/__init__.py
@@ -1,0 +1,6 @@
+"""SynthChat Parallel - Thread-pool based parallel generation workers.
+
+Location: SynthChat/parallel/__init__.py
+Purpose: Subpackage for parallel generation using ThreadPoolExecutor.
+Usage: from SynthChat.parallel.workers import run_parallel_generation
+"""
