+++ SynthChat/run.py
@@ -1,7 +1,10 @@
 """SynthChat CLI - Unified entry point for dataset generation and improvement.
 
 Location: SynthChat/run.py
-Purpose: Single CLI for both "generate" and "improve" modes
+Purpose: Single CLI for both "generate", "improve", and "validate" modes.
+         Contains only the argument parser, factory functions for LLM clients
+         and environment validators, and the main() dispatcher. Mode logic
+         lives in SynthChat.modes.{generate,improve,validate}.
 Usage: python -m SynthChat.run [generate|improve|validate] [options]
 
 Commands:
@@ -12,90 +15,16 @@ Commands:
 
 import argparse
 import sys
-import json
 from pathlib import Path
-from datetime import datetime, timezone
-from typing import Any, Dict, List, Optional
-from concurrent.futures import ThreadPoolExecutor, as_completed
-import threading
+from typing import Any, Dict, Optional
 
 from shared.llm import create_client
 from shared.environments import EnvironmentValidator
 from .utils.yaml_loader import load_yaml
-from .utils.docs_loader import DocsLoader, DocFile
-from .utils.logger import get_logger
-from .engine import ImprovementEngine
-from .generator import SynthChatGenerator, ScenarioLoader, _extract_shared_seed_spec
 
-
-class StreamingResultWriter:
-    """Writes generation results to JSONL incrementally as they complete.
-
-    Location: SynthChat/run.py
-    Purpose: Prevents data loss during long generation runs by streaming each
-    result to disk immediately instead of accumulating in memory.
-    Thread-safe for use with parallel workers via threading.Lock.
-
-    Usage:
-        with StreamingResultWriter(output_file, settings) as writer:
-            writer.write(result)  # Called after each example completes
-    """
-
-    def __init__(self, output_file: Path, settings: Dict):
-        self._output_file = output_file
-        self._settings = settings
-        self._lock = threading.Lock()
-        self._file = None
-        self._count = 0
-
-    def __enter__(self):
-        self._output_file.parent.mkdir(parents=True, exist_ok=True)
-        self._file = open(self._output_file, "w")
-
-        # Write metadata header as placeholder (will be updated at close)
-        if self._settings["output"]["include_metadata"]:
-            metadata = {
-                "_meta": {
-                    "synthchat_version": "1.0.0",
-                    "started_at": datetime.now(timezone.utc).isoformat(),
-                    "streaming": True
-                }
-            }
-            self._file.write(json.dumps(metadata) + "\n")
-            self._file.flush()
-
-        return self
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        if self._file and not self._file.closed:
-            self._file.flush()
-            self._file.close()
-        return False
-
-    def write(self, result) -> bool:
-        """Write a single GenerationResult to the output file.
-
-        Args:
-            result: GenerationResult with .example dict to serialize.
-
-        Returns:
-            True if write succeeded, False on I/O error.
-        """
-        try:
-            line = json.dumps(result.example) + "\n"
-            with self._lock:
-                self._file.write(line)
-                self._file.flush()
-                self._count += 1
-            return True
-        except (IOError, OSError) as e:
-            print(f"\nError writing result to {self._output_file}: {e}")
-            return False
-
-    @property
-    def count(self) -> int:
-        """Number of results successfully written."""
-        return self._count
+from .modes.generate import generate_mode
+from .modes.improve import improve_mode
+from .modes.validate import validate_mode
 
 
 def load_settings(config_dir: Path) -> Dict:
@@ -108,17 +37,16 @@ def load_settings(config_dir: Path) -> Dict:
 
 def create_llm_client(config: Dict, mode: str = "generation",
                       provider_override: str = None, model_override: str = None):
-    """
-    Create LLM client for generation or improvement.
+    """Create LLM client for generation or improvement.
 
     Args:
-        config: Settings configuration
-        mode: "generation" or "improvement"
-        provider_override: CLI override for provider (optional)
-        model_override: CLI override for model (optional)
+        config: Settings configuration.
+        mode: "generation" or "improvement".
+        provider_override: CLI override for provider (optional).
+        model_override: CLI override for model (optional).
 
     Returns:
-        LLM client from shared.llm
+        LLM client from shared.llm.
     """
     llm_config = config["llm"].get(mode, {})
     provider = provider_override or llm_config.get("provider", "lmstudio")
@@ -183,787 +111,6 @@ def create_environment_validator(settings: Dict, args) -> Optional[EnvironmentVa
         tool_schema_path=tool_schema_path,
         execution_config_path=execution_config_path,
     )
-
-
-def generate_mode(args):
-    """
-    Generate new dataset from scenarios.
-
-    Flow:
-        1. Load settings and scenarios
-        2. Create LLM clients (generation + improvement)
-        3. Create generator with improvement engine
-        4. Generate examples
-        5. Save results
-    """
-    print("=== SynthChat: Generate Mode ===\n")
-
-    # Load configuration
-    config_dir = Path(args.config_dir or "SynthChat/config")
-    settings = load_settings(config_dir)
-    logger = get_logger("synthchat_generate")
-
-    scenarios_dir = Path(args.scenarios_dir or "SynthChat/scenarios")
-    rubrics_dir = Path(args.rubrics_dir or "SynthChat/rubrics")
-
-    # Create LLM clients (CLI args override settings.yaml)
-    print("Initializing LLM clients...")
-    gen_client = create_llm_client(settings, mode="generation",
-                                   provider_override=args.provider, model_override=args.model)
-    improve_client = create_llm_client(settings, mode="improvement",
-                                       provider_override=args.provider, model_override=args.model)
-    environment_validator = create_environment_validator(settings, args)
-    if environment_validator is not None:
-        print(f"Environment validation enabled (backend={environment_validator.backend})")
-
-    # Create improvement engine
-    validation_config = config_dir / "validation.yaml"
-    engine = ImprovementEngine(
-        llm_client=improve_client,
-        rubrics_dir=rubrics_dir,
-        config_path=validation_config,
-        logger=logger,
-        enable_interactions=settings["logging"]["save_interactions"]
-    )
-
-    # Create generator
-    generator = SynthChatGenerator(
-        config_dir=config_dir,
-        scenarios_dir=scenarios_dir,
-        rubrics_dir=rubrics_dir,
-        llm_client=gen_client,
-        engine=engine,
-        environment_validator=environment_validator,
-        enable_stage_validation=settings["generation"]["stage_validation"],
-        logger=logger,
-    )
-
-    # Load targets
-    if args.targets_file:
-        with open(args.targets_file) as f:
-            targets = json.load(f)
-    else:
-        # Use default targets from settings
-        targets = settings["defaults"]["targets"]
-
-    shared_seed_spec, targets = _extract_shared_seed_spec(targets)
-
-    # Filter targets if specific scenarios requested
-    if args.scenarios:
-        targets = {k: v for k, v in targets.items() if k in args.scenarios}
-        if shared_seed_spec and shared_seed_spec.get("targets"):
-            shared_seed_spec = {
-                **shared_seed_spec,
-                "targets": [key for key in shared_seed_spec["targets"] if key in targets],
-            }
-
-    print(f"\nGeneration targets:")
-    total_examples = 0
-    if shared_seed_spec:
-        shared_targets = set(shared_seed_spec.get("targets") or targets.keys())
-        print(
-            f"  [shared seed] scenario={shared_seed_spec['scenario']} "
-            f"seed_count={shared_seed_spec['seed_count']} targets={len(shared_targets)}"
-        )
-    for scenario_key, raw_target in targets.items():
-        target_spec = _normalize_target_spec(raw_target)
-        if shared_seed_spec and (not shared_seed_spec.get("targets") or scenario_key in set(shared_seed_spec.get("targets") or [])):
-            scenario_total = shared_seed_spec["seed_count"] * target_spec["rollouts_per_seed"]
-        else:
-            scenario_total = target_spec["seed_count"] * target_spec["rollouts_per_seed"]
-        total_examples += scenario_total
-        if target_spec["rollouts_per_seed"] == 1:
-            print(f"  {scenario_key}: {scenario_total}")
-        else:
-            print(
-                f"  {scenario_key}: {target_spec['seed_count']} seed(s) x "
-                f"{target_spec['rollouts_per_seed']} rollout(s) = {scenario_total}"
-            )
-    print(f"Total: {total_examples} examples\n")
-
-    # Load docs if provided
-    docs: List[DocFile] = []
-    if args.docs:
-        print(f"Loading docs from: {args.docs}")
-        docs = DocsLoader().load(args.docs)
-        print(f"Loaded {len(docs)} document(s)")
-        print(f"Will generate {args.per_doc} example(s) per doc\n")
-
-    # Generate
-    max_iterations = args.max_iterations or settings["improvement"]["max_iterations"]
-    args.workers = max(1, args.workers)
-    num_workers = args.workers
-    results = []
-
-    # Determine output path early so we can stream results to disk
-    output_file = Path(args.output) if args.output else _generate_output_path(settings)
-
-    with StreamingResultWriter(output_file, settings) as writer:
-        if docs and num_workers > 1:
-            # Parallel docs-based generation with multiple workers
-            print(f"Using {num_workers} parallel workers for {len(docs)} doc(s)\n")
-
-            # Build work items: one per (doc, repetition, scenario, count) combination
-            work_items = []
-            task_id = 0
-            for doc in docs:
-                for rep in range(args.per_doc):
-                    shared_targets = set((shared_seed_spec or {}).get("targets") or [])
-                    if shared_seed_spec:
-                        shared_scenario = generator.scenario_loader.get_scenario(shared_seed_spec["scenario"])
-                        if not shared_scenario:
-                            raise ValueError(f"Shared seed scenario not found: {shared_seed_spec['scenario']}")
-                        for shared_seed_index in range(shared_seed_spec["seed_count"]):
-                            shared_seed_id = f"{shared_seed_spec['scenario']}:shared_seed:{shared_seed_index + 1}"
-                            shared_seed_bundle = generator.prepare_seed_bundle(
-                                scenario_key=shared_seed_spec["scenario"],
-                                seed_id=shared_seed_id,
-                                scenario=shared_scenario,
-                                randomize_params=True,
-                                doc_context=doc,
-                            )
-                            for scenario_key, raw_target in targets.items():
-                                if shared_targets and scenario_key not in shared_targets:
-                                    continue
-                                scenario = generator.scenario_loader.get_scenario(scenario_key)
-                                if not scenario:
-                                    print(f"Warning: Scenario not found: {scenario_key}")
-                                    continue
-                                target_spec = _normalize_target_spec(raw_target)
-                                for rollout_index in range(target_spec["rollouts_per_seed"]):
-                                    work_items.append((
-                                        scenario_key, scenario, max_iterations,
-                                        config_dir, scenarios_dir, rubrics_dir,
-                                        settings, args.provider, args.model, doc, task_id,
-                                        _serialize_environment_options(environment_validator),
-                                        shared_seed_bundle,
-                                        {
-                                            "seed_id": shared_seed_id,
-                                            "seed_index": shared_seed_index,
-                                            "seed_count": shared_seed_spec["seed_count"],
-                                            "rollout_index": rollout_index,
-                                            "rollouts_per_seed": target_spec["rollouts_per_seed"],
-                                            "shared_seed_source": shared_seed_spec["scenario"],
-                                            "shared_across_scenarios": True,
-                                        },
-                                    ))
-                                    task_id += 1
-                    for scenario_key, raw_target in targets.items():
-                        if shared_seed_spec and (not shared_seed_spec.get("targets") or scenario_key in set(shared_seed_spec.get("targets") or [])):
-                            continue
-                        scenario = generator.scenario_loader.get_scenario(scenario_key)
-                        if not scenario:
-                            print(f"Warning: Scenario not found: {scenario_key}")
-                            continue
-                        target_spec = _normalize_target_spec(raw_target)
-                        for seed_index in range(target_spec["seed_count"]):
-                            seed_id = f"{scenario_key}:seed:{seed_index + 1}"
-                            seed_bundle = generator.prepare_seed_bundle(
-                                scenario_key=scenario_key,
-                                seed_id=seed_id,
-                                scenario=scenario,
-                                randomize_params=True,
-                                doc_context=doc,
-                            )
-                            for rollout_index in range(target_spec["rollouts_per_seed"]):
-                                work_items.append((
-                                    scenario_key, scenario, max_iterations,
-                                    config_dir, scenarios_dir, rubrics_dir,
-                                    settings, args.provider, args.model, doc, task_id,
-                                    _serialize_environment_options(environment_validator),
-                                    seed_bundle,
-                                    {
-                                        "seed_id": seed_id,
-                                        "seed_index": seed_index,
-                                        "seed_count": target_spec["seed_count"],
-                                        "rollout_index": rollout_index,
-                                        "rollouts_per_seed": target_spec["rollouts_per_seed"],
-                                    },
-                                ))
-                                task_id += 1
-
-            results.extend(_run_parallel_generation(work_items, num_workers, writer))
-        elif docs:
-            # Sequential docs-based generation (single worker)
-            total_docs = len(docs)
-            for doc_idx, doc in enumerate(docs, 1):
-                print(f"\n--- Document {doc_idx}/{total_docs}: {doc.path} ---")
-                for rep in range(args.per_doc):
-                    if args.per_doc > 1:
-                        print(f"  Repetition {rep + 1}/{args.per_doc}")
-                    batch_results = generator.generate_batch(
-                        targets=targets,
-                        max_iterations=max_iterations,
-                        randomize_params=True,
-                        doc_context=doc,
-                        on_result=writer.write,
-                        shared_seed_spec=shared_seed_spec,
-                    )
-                    results.extend(batch_results)
-        elif num_workers > 1:
-            # Parallel generation with multiple workers (no docs)
-            print(f"Using {num_workers} parallel workers\n")
-
-            # Build work items
-            work_items = []
-            task_id = 0
-            shared_targets = set((shared_seed_spec or {}).get("targets") or [])
-            if shared_seed_spec:
-                shared_scenario = generator.scenario_loader.get_scenario(shared_seed_spec["scenario"])
-                if not shared_scenario:
-                    raise ValueError(f"Shared seed scenario not found: {shared_seed_spec['scenario']}")
-                for shared_seed_index in range(shared_seed_spec["seed_count"]):
-                    shared_seed_id = f"{shared_seed_spec['scenario']}:shared_seed:{shared_seed_index + 1}"
-                    shared_seed_bundle = generator.prepare_seed_bundle(
-                        scenario_key=shared_seed_spec["scenario"],
-                        seed_id=shared_seed_id,
-                        scenario=shared_scenario,
-                        randomize_params=True,
-                    )
-                    for scenario_key, raw_target in targets.items():
-                        if shared_targets and scenario_key not in shared_targets:
-                            continue
-                        scenario = generator.scenario_loader.get_scenario(scenario_key)
-                        if not scenario:
-                            print(f"Warning: Scenario not found: {scenario_key}")
-                            continue
-                        target_spec = _normalize_target_spec(raw_target)
-                        for rollout_index in range(target_spec["rollouts_per_seed"]):
-                            work_items.append((
-                                scenario_key, scenario, max_iterations,
-                                config_dir, scenarios_dir, rubrics_dir,
-                                settings, args.provider, args.model, None, task_id,
-                                _serialize_environment_options(environment_validator),
-                                shared_seed_bundle,
-                                {
-                                    "seed_id": shared_seed_id,
-                                    "seed_index": shared_seed_index,
-                                    "seed_count": shared_seed_spec["seed_count"],
-                                    "rollout_index": rollout_index,
-                                    "rollouts_per_seed": target_spec["rollouts_per_seed"],
-                                    "shared_seed_source": shared_seed_spec["scenario"],
-                                    "shared_across_scenarios": True,
-                                },
-                            ))
-                            task_id += 1
-            for scenario_key, raw_target in targets.items():
-                if shared_seed_spec and (not shared_seed_spec.get("targets") or scenario_key in set(shared_seed_spec.get("targets") or [])):
-                    continue
-                scenario = generator.scenario_loader.get_scenario(scenario_key)
-                if not scenario:
-                    print(f"Warning: Scenario not found: {scenario_key}")
-                    continue
-                target_spec = _normalize_target_spec(raw_target)
-                for seed_index in range(target_spec["seed_count"]):
-                    seed_id = f"{scenario_key}:seed:{seed_index + 1}"
-                    seed_bundle = generator.prepare_seed_bundle(
-                        scenario_key=scenario_key,
-                        seed_id=seed_id,
-                        scenario=scenario,
-                        randomize_params=True,
-                    )
-                    for rollout_index in range(target_spec["rollouts_per_seed"]):
-                        work_items.append((
-                            scenario_key, scenario, max_iterations,
-                            config_dir, scenarios_dir, rubrics_dir,
-                            settings, args.provider, args.model, None, task_id,
-                            _serialize_environment_options(environment_validator),
-                            seed_bundle,
-                            {
-                                "seed_id": seed_id,
-                                "seed_index": seed_index,
-                                "seed_count": target_spec["seed_count"],
-                                "rollout_index": rollout_index,
-                                "rollouts_per_seed": target_spec["rollouts_per_seed"],
-                            },
-                        ))
-                        task_id += 1
-
-            results.extend(_run_parallel_generation(work_items, num_workers, writer))
-        else:
-            # Standard sequential generation (no docs)
-            results = generator.generate_batch(
-                targets=targets,
-                max_iterations=max_iterations,
-                randomize_params=True,
-                on_result=writer.write,
-                shared_seed_spec=shared_seed_spec,
-            )
-
-        print(f"\nStreamed {writer.count} examples to {output_file}")
-
-    # Print summary
-    _print_summary(results, output_file)
-
-
-def improve_mode(args):
-    """
-    Improve existing dataset with rubrics.
-
-    Flow:
-        1. Load settings and dataset
-        2. Create improvement LLM client
-        3. Create improvement engine
-        4. Process each example
-        5. Save improved dataset
-    """
-    print("=== SynthChat: Improve Mode ===\n")
-
-    # Load configuration
-    config_dir = Path(args.config_dir or "SynthChat/config")
-    settings = load_settings(config_dir)
-    rubrics_dir = Path(args.rubrics_dir or "SynthChat/rubrics")
-
-    # Load input dataset
-    if not args.input:
-        print("Error: --input required for improve mode")
-        sys.exit(1)
-
-    input_path = Path(args.input)
-    if not input_path.exists():
-        print(f"Error: Input file not found: {input_path}")
-        sys.exit(1)
-
-    # Load examples
-    examples = []
-    with open(input_path) as f:
-        for line in f:
-            line = line.strip()
-            if line:
-                examples.append(json.loads(line))
-
-    # Apply line range if specified
-    if args.start_line or args.end_line:
-        start = (args.start_line or 1) - 1  # Convert to 0-indexed
-        end = args.end_line or len(examples)
-        examples = examples[start:end]
-        print(f"Processing lines {start + 1} to {end}")
-
-    print(f"Loaded {len(examples)} examples from {input_path}\n")
-
-    # Create LLM client (CLI args override settings.yaml)
-    print("Initializing improvement LLM client...")
-    improve_client = create_llm_client(settings, mode="improvement",
-                                       provider_override=args.provider, model_override=args.model)
-
-    # Create improvement engine
-    validation_config = config_dir / "validation.yaml"
-    engine = ImprovementEngine(
-        llm_client=improve_client,
-        rubrics_dir=rubrics_dir,
-        config_path=validation_config,
-        enable_interactions=settings["logging"]["save_interactions"]
-    )
-
-    # Determine rubrics to use
-    rubrics = args.rubrics or settings["improvement"]["default_rubrics"]
-    if isinstance(rubrics, str):
-        rubrics = [r.strip() for r in rubrics.split(",")]
-
-    print(f"Applying rubrics: {', '.join(rubrics)}\n")
-
-    # Process examples
-    max_iterations = args.max_iterations or settings["improvement"]["max_iterations"]
-    improved_examples = []
-    passed_count = 0
-    failed_count = 0
-
-    for i, example in enumerate(examples, 1):
-        print(f"Processing {i}/{len(examples)}...")
-
-        try:
-            result = engine.run(
-                example=example,
-                rubric_keys=rubrics,
-                max_iterations=max_iterations
-            )
-
-            improved_examples.append(result.improved_example)
-            if result.passed:
-                passed_count += 1
-            else:
-                failed_count += 1
-
-        except Exception as e:
-            print(f"  Error: {e}")
-            improved_examples.append(example)  # Keep original
-            failed_count += 1
-
-    # Save results
-    output_file = args.output or _generate_output_path(settings, input_path)
-    with open(output_file, "w") as f:
-        for example in improved_examples:
-            f.write(json.dumps(example) + "\n")
-
-    # Print summary
-    print(f"\n=== Summary ===")
-    print(f"Total processed: {len(examples)}")
-    print(f"Passed: {passed_count}")
-    print(f"Failed: {failed_count}")
-    print(f"Output: {output_file}")
-
-
-def validate_mode(args):
-    """
-    Validate dataset against rubrics without improving.
-
-    Flow:
-        1. Load dataset and settings
-        2. Create engine (read-only mode)
-        3. Run validation on each example
-        4. Report results
-    """
-    print("=== SynthChat: Validate Mode ===\n")
-
-    # Load configuration
-    config_dir = Path(args.config_dir or "SynthChat/config")
-    settings = load_settings(config_dir)
-    rubrics_dir = Path(args.rubrics_dir or "SynthChat/rubrics")
-
-    # Load input dataset
-    if not args.input:
-        print("Error: --input required for validate mode")
-        sys.exit(1)
-
-    input_path = Path(args.input)
-    if not input_path.exists():
-        print(f"Error: Input file not found: {input_path}")
-        sys.exit(1)
-
-    # Load examples
-    examples = []
-    with open(input_path) as f:
-        for line_num, line in enumerate(f, 1):
-            line = line.strip()
-            if line:
-                try:
-                    examples.append((line_num, json.loads(line)))
-                except json.JSONDecodeError as e:
-                    print(f"Warning: Skipping malformed JSON at line {line_num}: {e}")
-
-    print(f"Loaded {len(examples)} examples from {input_path}\n")
-
-    # Create LLM client for judging (CLI args override settings.yaml)
-    print("Initializing validation LLM client...")
-    validate_client = create_llm_client(settings, mode="improvement",
-                                        provider_override=args.provider, model_override=args.model)
-
-    # Create improvement engine
-    validation_config = config_dir / "validation.yaml"
-    engine = ImprovementEngine(
-        llm_client=validate_client,
-        rubrics_dir=rubrics_dir,
-        config_path=validation_config,
-        enable_interactions=False  # No logging for validation-only
-    )
-
-    # Determine rubrics to use
-    rubrics = args.rubrics or settings["improvement"]["default_rubrics"]
-    if isinstance(rubrics, str):
-        rubrics = [r.strip() for r in rubrics.split(",")]
-
-    print(f"Validating against rubrics: {', '.join(rubrics)}\n")
-
-    # Validate examples (max_iterations=1 means just judge, no improvement)
-    total_passed = 0
-    total_failed = 0
-    failures_by_rubric = {rubric: [] for rubric in rubrics}
-    failing_lines = []
-
-    for line_num, example in examples:
-        try:
-            # Run validation with max_iterations=1 (judge only, no improvement loop)
-            result = engine.run(
-                example=example,
-                rubric_keys=rubrics,
-                max_iterations=1  # Just judge once, don't improve
-            )
-
-            if result.passed:
-                total_passed += 1
-            else:
-                total_failed += 1
-                failing_lines.append(line_num)
-
-                # Track which rubrics failed
-                for rubric in rubrics:
-                    score = result.final_scores.get(rubric, 0.0)
-                    # Load rubric to get threshold
-                    rubric_config = engine.rubric_repo.get_rubric(rubric)
-                    threshold = rubric_config.get("pass_threshold", 0.8)
-                    if score < threshold:
-                        failures_by_rubric[rubric].append(line_num)
-
-        except Exception as e:
-            print(f"  Error validating line {line_num}: {e}")
-            total_failed += 1
-            failing_lines.append(line_num)
-
-    # Print summary
-    total = len(examples)
-    pass_rate = (total_passed / total * 100) if total > 0 else 0
-
-    print(f"\n=== Validation Summary ===")
-    print(f"Total examples: {total}")
-    print(f"Passed: {total_passed} ({pass_rate:.1f}%)")
-    print(f"Failed: {total_failed} ({100 - pass_rate:.1f}%)")
-
-    if failing_lines:
-        print(f"\nFailing lines: {', '.join(map(str, failing_lines[:20]))}")
-        if len(failing_lines) > 20:
-            print(f"  ... and {len(failing_lines) - 20} more")
-
-    print(f"\n=== Failures by Rubric ===")
-    for rubric, failed_lines in failures_by_rubric.items():
-        if failed_lines:
-            print(f"{rubric}: {len(failed_lines)} failures")
-            print(f"  Lines: {', '.join(map(str, failed_lines[:10]))}")
-            if len(failed_lines) > 10:
-                print(f"  ... and {len(failed_lines) - 10} more")
-
-    if total_failed > 0:
-        print(f"\nRun with 'improve' mode to fix failing examples:")
-        print(f"  python -m SynthChat.run improve --input {input_path} --rubrics {','.join(rubrics)}")
-
-
-def _run_parallel_generation(work_items: List, num_workers: int,
-                             writer=None) -> List:
-    """
-    Execute work items in parallel using a thread pool.
-
-    Shared execution logic for both docs-based and non-docs parallel generation.
-    Handles progress tracking, error reporting, streaming to disk, and graceful
-    shutdown on interrupts.
-
-    Args:
-        work_items: List of tuples to pass to _generate_single_example
-        num_workers: Number of parallel worker threads
-        writer: Optional StreamingResultWriter to stream results as they complete
-
-    Returns:
-        List of GenerationResult objects (successful results only),
-        sorted by task_id to preserve input ordering.
-    """
-    if not work_items:
-        print("No work items to process (check scenario names)")
-        return []
-
-    total = len(work_items)
-    completed = 0
-    lock = threading.Lock()
-    indexed_results = []  # List of (task_id, result) for ordering
-
-    def update_progress():
-        nonlocal completed
-        with lock:
-            completed += 1
-            pct = (completed / total * 100) if total > 0 else 0
-            print(f"\rProgress: {completed}/{total} ({pct:.1f}%)", end="", flush=True)
-
-    with ThreadPoolExecutor(max_workers=num_workers) as executor:
-        futures = {executor.submit(_generate_single_example, item): item for item in work_items}
-
-        try:
-            for future in as_completed(futures):
-                result, error, task_id = future.result()
-                if error:
-                    print(f"\n{error}")
-                if result:
-                    if writer:
-                        writer.write(result)
-                    indexed_results.append((task_id, result))
-                update_progress()
-        except BaseException as e:
-            print(f"\nInterrupted: {e}. Shutting down workers...")
-            executor.shutdown(wait=False, cancel_futures=True)
-            raise
-
-    print()  # Newline after progress
-
-    # Sort by task_id to preserve input document order
-    indexed_results.sort(key=lambda x: x[0])
-    return [result for _, result in indexed_results]
-
-
-def _serialize_environment_options(environment_validator: Optional[EnvironmentValidator]) -> Dict:
-    """Serialize environment validator configuration for worker fan-out."""
-    if environment_validator is None:
-        return {"backend": "none"}
-    return {
-        "backend": environment_validator.backend,
-        "template": environment_validator.e2b_template,
-        "api_key": environment_validator.e2b_api_key,
-        "timeout_seconds": environment_validator.timeout_seconds,
-        "tool_schema_path": str(environment_validator.tool_schema_path),
-        "execution_config_path": str(environment_validator.execution_config_path),
-    }
-
-
-def _create_environment_validator_from_options(options: Dict) -> Optional[EnvironmentValidator]:
-    """Recreate environment validator from serialized options."""
-    backend = (options or {}).get("backend", "none")
-    if backend == "none":
-        return None
-    return EnvironmentValidator(
-        backend=backend,
-        e2b_template=(options or {}).get("template"),
-        e2b_api_key=(options or {}).get("api_key"),
-        timeout_seconds=float((options or {}).get("timeout_seconds", 120.0)),
-        tool_schema_path=(options or {}).get("tool_schema_path"),
-        execution_config_path=(options or {}).get("execution_config_path"),
-    )
-
-
-def _create_worker_generator(config_dir: Path, scenarios_dir: Path, rubrics_dir: Path,
-                              settings: Dict, provider: str = None, model: str = None,
-                              environment_options: Optional[Dict] = None):
-    """Create a new generator instance for a worker thread."""
-    logger = get_logger("synthchat_generate_worker")
-    gen_client = create_llm_client(settings, mode="generation",
-                                   provider_override=provider, model_override=model)
-    improve_client = create_llm_client(settings, mode="improvement",
-                                       provider_override=provider, model_override=model)
-
-    validation_config = config_dir / "validation.yaml"
-    engine = ImprovementEngine(
-        llm_client=improve_client,
-        rubrics_dir=rubrics_dir,
-        config_path=validation_config,
-        logger=logger,
-        enable_interactions=settings["logging"]["save_interactions"]
-    )
-
-    generator = SynthChatGenerator(
-        config_dir=config_dir,
-        scenarios_dir=scenarios_dir,
-        rubrics_dir=rubrics_dir,
-        llm_client=gen_client,
-        engine=engine,
-        environment_validator=_create_environment_validator_from_options(environment_options or {}),
-        enable_stage_validation=settings["generation"]["stage_validation"],
-        logger=logger,
-    )
-    return generator
-
-
-def _generate_single_example(args_tuple):
-    """Worker function to generate a single example.
-
-    Returns:
-        Tuple of (result, error, task_id) where task_id preserves input ordering.
-    """
-    (scenario_key, scenario, max_iterations, config_dir, scenarios_dir,
-     rubrics_dir, settings, provider, model, doc_context, task_id,
-     environment_options, seed_bundle, rollout_metadata) = args_tuple
-
-    try:
-        # Each worker creates its own generator (thread-safe LLM clients)
-        generator = _create_worker_generator(
-            config_dir, scenarios_dir, rubrics_dir, settings, provider, model,
-            environment_options=environment_options
-        )
-
-        result = generator.generate_single(
-            scenario_key,
-            scenario,
-            max_iterations,
-            True,
-            doc_context,
-            seed_bundle=seed_bundle,
-            rollout_metadata=rollout_metadata,
-        )
-        return result, None, task_id
-    except Exception as e:
-        return None, f"Task {task_id} error for {scenario_key}: {e}", task_id
-
-
-def _normalize_target_spec(raw_target: Any) -> Dict[str, int]:
-    if isinstance(raw_target, bool):
-        raise ValueError("Boolean target specs are not supported")
-    if isinstance(raw_target, int):
-        if raw_target < 0:
-            raise ValueError("Target counts must be non-negative")
-        return {"seed_count": raw_target, "rollouts_per_seed": 1}
-    if not isinstance(raw_target, dict):
-        raise ValueError(f"Unsupported target spec: {raw_target!r}")
-
-    count = raw_target.get("count")
-    seed_count = raw_target.get("seed_count", count if count is not None else 1)
-    rollouts_per_seed = raw_target.get("rollouts_per_seed", 1)
-
-    seed_count = int(seed_count)
-    rollouts_per_seed = int(rollouts_per_seed)
-    if seed_count < 0 or rollouts_per_seed < 0:
-        raise ValueError("Target specs must use non-negative integers")
-    return {"seed_count": seed_count, "rollouts_per_seed": rollouts_per_seed}
-
-
-def _generate_output_path(settings: Dict, input_path: Optional[Path] = None) -> Path:
-    """
-    Generate output file path with datetime versioning.
-
-    Args:
-        settings: Settings configuration
-        input_path: Input file path (for improve mode)
-
-    Returns:
-        Output file path
-    """
-    timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
-
-    if input_path:
-        # Improve mode: append timestamp to input name
-        stem = input_path.stem
-        # Remove existing version if present (e.g., _v1.8)
-        if "_v" in stem:
-            stem = stem.split("_v")[0]
-        return input_path.parent / f"{stem}_{timestamp}.jsonl"
-    else:
-        # Generate mode: use default directory
-        default_dir = Path(settings["output"]["default_dir"])
-        default_dir.mkdir(parents=True, exist_ok=True)
-        return default_dir / f"synthchat_{timestamp}.jsonl"
-
-
-def _save_results(results: List, output_file: Path, settings: Dict):
-    """Save generation results to JSONL with metadata."""
-    output_file.parent.mkdir(parents=True, exist_ok=True)
-
-    with open(output_file, "w") as f:
-        # Write metadata header if enabled
-        if settings["output"]["include_metadata"]:
-            metadata = {
-                "_meta": {
-                    "synthchat_version": "1.0.0",
-                    "generated_at": datetime.utcnow().isoformat(),
-                    "stats": {
-                        "total": len(results),
-                        "passed": sum(1 for r in results if r.success),
-                        "failed": sum(1 for r in results if not r.success),
-                        "avg_iterations": sum(r.iterations for r in results) / len(results) if results else 0
-                    }
-                }
-            }
-            f.write(json.dumps(metadata) + "\n")
-
-        # Write examples
-        for result in results:
-            f.write(json.dumps(result.example) + "\n")
-
-
-def _print_summary(results: List, output_file: Path):
-    """Print generation summary."""
-    total = len(results)
-    passed = sum(1 for r in results if r.success)
-    failed = total - passed
-    avg_iterations = sum(r.iterations for r in results) / total if total else 0
-
-    print(f"\n=== Summary ===")
-    print(f"Total generated: {total}")
-    print(f"Passed: {passed} ({passed / total * 100:.1f}%)")
-    print(f"Failed: {failed} ({failed / total * 100:.1f}%)")
-    print(f"Avg iterations: {avg_iterations:.1f}")
-    print(f"Output: {output_file}")
 
 
 def main():
@@ -1047,13 +194,26 @@ def main():
         parser.print_help()
         sys.exit(1)
 
-    # Route to appropriate mode
+    # Route to appropriate mode, injecting factory functions
     if args.command == "generate":
-        generate_mode(args)
+        generate_mode(
+            args,
+            load_settings=load_settings,
+            create_llm_client=create_llm_client,
+            create_environment_validator=create_environment_validator,
+        )
     elif args.command == "improve":
-        improve_mode(args)
+        improve_mode(
+            args,
+            load_settings=load_settings,
+            create_llm_client=create_llm_client,
+        )
     elif args.command == "validate":
-        validate_mode(args)
+        validate_mode(
+            args,
+            load_settings=load_settings,
+            create_llm_client=create_llm_client,
+        )
 
 
 if __name__ == "__main__":
