+++ SynthChat/schemas/__init__.py
@@ -1,0 +1,8 @@
+"""SynthChat Schemas - JSON schema builders for environment and tool response generation.
+
+Location: SynthChat/schemas/__init__.py
+Purpose: Subpackage for constructing JSON schemas used during synthetic data generation,
+         including canonical environment schemas and use_tools response schemas.
+Usage: from SynthChat.schemas.environment_schema import _build_canonical_environment_schema
+       from SynthChat.schemas.tool_response_schema import _build_use_tools_response_schema
+"""
